Move instance list to separate file
